requirements.txt
altgraph==0.17
aubio @ file:///D:/bld/aubio_1611074891327/work
backcall==0.2.0
beautifulsoup4==4.9.3
boto3==1.16.40
botocore==1.19.40
bs4==0.0.1
cachetools==4.2.0
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
colorama==0.4.4
cryptography==3.4.5
cycler==0.10.0
decorator==4.4.2
fake-useragent==0.1.11
future==0.18.2
futures==3.1.1
google-api-core==1.24.1
google-api-python-client==1.12.8
google-auth==1.24.0
google-auth-httplib2==0.0.4
Google-Images-Search==1.3.4
googleapis-common-protos==1.52.0
goslate==1.5.1
httplib2==0.18.1
idna==2.10
imageio==2.9.0
imageio-ffmpeg==0.4.3
ipython==7.20.0
ipython-genutils==0.2.0
jedi==0.18.0
jmespath==0.10.0
joblib==1.0.0
kiwisolver==1.3.1
language-check==1.1
lxml==4.6.2
markovify==0.8.3
matplotlib==3.3.3
moviepy==1.0.3
multitasking==0.0.9
mutagen==1.45.1
nltk==3.5
numpy @ file:///D:/bld/numpy_1612192921641/work
pandas==1.2.0
pandas-datareader==0.9.0
parso==0.8.1
pefile==2019.4.18
pickleshare==0.7.5
Pillow==7.1.2
playsound==1.2.2
proglog==0.1.9
prompt-toolkit==3.0.16
protobuf==3.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAudio @ file:///C:/Users/Jimone%20Treburtski/Downloads/PyAudio-0.2.11-cp38-cp38-win_amd64.whl
pycparser==2.20
PyDictionary==2.0.1
pydub==0.24.1
pyfiglet==0.8.post1
pygame==2.0.1
Pygments==2.8.0
pyimgur==0.6.0
pyinstaller==4.2
pyinstaller-hooks-contrib==2020.11
pyparsing==2.4.7
PySocks==1.7.1
python-dateutil==2.8.1
python-resize-image==1.1.19
pytz==2020.5
pywin32-ctypes==0.2.0
rake-nltk==1.0.4
regex==2020.11.13
requests==2.25.1
rsa==4.7
s3transfer==0.3.3
scipy==1.6.0
sigfig==1.1.8
six==1.15.0
sortedcontainers==2.3.0
soupsieve==2.1
termcolor==1.1.0
ticker==0.13
tk==0.1.0
torpy==1.1.5
tqdm==4.56.0
traitlets==5.0.5
Unidecode==1.1.2
uritemplate==3.0.1
urllib3==1.26.2
Wave==0.0.2
wcwidth==0.2.5
wget==3.2
wincertstore==0.2
windows-curses==2.2.0
yfinance==0.1.55
youtube-dl==2021.1.16